Renamed Helpers.h and Helpers.cpp to ImHelpers.h and ImHelpers.cpp because their names are too general and sometimes it conflicts with another addon (such as ofxCv).
