+++ src/ofxImGui.h
@@ -2,4 +2,4 @@
 
 #include "imgui.h"
 #include "Gui.h"
-#include "Helpers.h"
+#include "ImHelpers.h"
